+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # UNRELEASED
 
+IMPROVEMENETS
+
+* Added a flag to skip legacy duplicate telemetry. [GH-630](https://github.com/hashicorp/raft/pull/630)
+
 # 1.7.0 (June 5th, 2024)
 
 CHANGES